更新jar依赖

+++ src/main/java/com/syes/syes_springboot/controller/ItemController.java
@@ -58,7 +58,7 @@ public class ItemController {
     public Result slectByPage(@RequestParam("pagesize") int pagesize, @RequestParam("currentpage") int currentPage) {
 
         QueryWrapper wrapper = new QueryWrapper();
-        Integer total = itemMapper.selectCount(wrapper); //获取总数
+        Integer total = itemMapper.selectCount(wrapper).intValue(); //获取总数
         int StartPage = (currentPage - 1) * pagesize; //开始页数
         List<Item> itemList = itemMapper.slectByPage(StartPage, pagesize); //列表
         HashMap<String, Object> map = new HashMap<>();
