+++ src/main/java/com/syes/syes_springboot/controller/UserController.java
@@ -70,7 +70,7 @@ public class UserController {
     public Result slectByPage(@RequestParam("pagesize") int pagesize, @RequestParam("currentpage") int currentPage) {
 
         QueryWrapper wrapper = new QueryWrapper();
-        Integer total = userMapper.selectCount(wrapper); //获取总数
+        Integer total = userMapper.selectCount(wrapper).intValue(); //获取总数
         int StartPage = (currentPage - 1) * pagesize; //开始页数
         List<User> userList = userMapper.slectByPage(StartPage, pagesize); //列表
